pyproject.toml
[tool.poetry]
name = "tg-bot-demo"
version = "0.1.0"
description = ""
authors = ["Jay Hsia <jayhsia1997@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "*"
uvicorn = "*"
pydantic = "*"
pydantic-settings = "*"
python-dotenv = "*"
python-telegram-bot = "^20.6"
sentry-sdk = "^1.34.0"


[tool.poetry.group.dev.dependencies]
pytest = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
